06_Projects/cleanstat-autonomous-agent/requirements.txt
groq>=0.5.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
